+++ example/lib/main.dart
@@ -258,8 +258,7 @@ class StorageActions extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceAround,
+    return Wrap(
       children: <Widget>[
         ElevatedButton(
           child: const Text('read'),
@@ -278,12 +277,50 @@ class StorageActions extends StatelessWidget {
           },
         ),
         ElevatedButton(
+          child: const Text('read with force'),
+          onPressed: () async {
+            _logger.fine(
+                'reading with forceBiometricAuthentication from ${storageFile.name}');
+            try {
+              final result = await storageFile.read(
+                forceBiometricAuthentication: true,
+              );
+              _logger.fine('read: {$result}');
+            } on AuthException catch (e) {
+              if (e.code == AuthExceptionCode.userCanceled) {
+                _logger.info('User canceled.');
+                return;
+              }
+              rethrow;
+            }
+          },
+        ),
+        ElevatedButton(
           child: const Text('write'),
           onPressed: () async {
             _logger.fine('Going to write...');
             try {
               await storageFile
                   .write(' [${DateTime.now()}] ${writeController.text}');
+              _logger.info('Written content.');
+            } on AuthException catch (e) {
+              if (e.code == AuthExceptionCode.userCanceled) {
+                _logger.info('User canceled.');
+                return;
+              }
+              rethrow;
+            }
+          },
+        ),
+        ElevatedButton(
+          child: const Text('write with forceBiometricAuthentication'),
+          onPressed: () async {
+            _logger.fine('Going to write with force...');
+            try {
+              await storageFile.write(
+                ' [${DateTime.now()}] ${writeController.text}',
+                forceBiometricAuthentication: true,
+              );
               _logger.info('Written content.');
             } on AuthException catch (e) {
               if (e.code == AuthExceptionCode.userCanceled) {